moved Tasks into their own classes
