+++ src/main/java/net/fameless/forcebattle/game/ObjectiveManager.java
@@ -3,10 +3,10 @@ package net.fameless.forcebattle.game;
 import net.fameless.forcebattle.ForceBattle;
 import net.fameless.forcebattle.configuration.SettingsManager;
 import net.fameless.forcebattle.player.BattlePlayer;
-import net.fameless.forcebattle.util.Advancement;
+import net.fameless.forcebattle.game.data.Advancement;
 import net.fameless.forcebattle.util.BattleType;
 import net.fameless.forcebattle.util.BukkitUtil;
-import net.fameless.forcebattle.util.Structure;
+import net.fameless.forcebattle.game.data.Structure;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
