+++ src/main/java/net/fameless/forcebattle/game/tasks/CoordsTask.java
@@ -1,0 +1,59 @@
+package net.fameless.forcebattle.game.tasks;
+
+import net.fameless.forcebattle.ForceBattle;
+import net.fameless.forcebattle.caption.Caption;
+import net.fameless.forcebattle.configuration.SettingsManager;
+import net.fameless.forcebattle.player.BattlePlayer;
+import net.fameless.forcebattle.util.BattleType;
+import net.fameless.forcebattle.util.BukkitUtil;
+import net.fameless.forcebattle.util.Format;
+import net.fameless.forcebattle.util.Toast;
+import net.kyori.adventure.text.Component;
+import net.kyori.adventure.text.minimessage.tag.Tag;
+import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver;
+import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
+import org.bukkit.Location;
+import org.bukkit.Material;
+import org.bukkit.entity.Player;
+
+public class CoordsTask implements ForceTask {
+
+    @Override
+    public void runTick() {
+        if (!SettingsManager.isEnabled(SettingsManager.Setting.FORCE_COORDS)) return;
+        if (!ForceBattle.getTimer().isRunning()) return;
+
+        for (Player player : Bukkit.getOnlinePlayers()) {
+            BattlePlayer battlePlayer = BattlePlayer.adapt(player);
+            if (battlePlayer.isExcluded()) continue;
+
+            String objectiveString = battlePlayer.getObjective().getObjectiveString();
+
+            if (BukkitUtil.convertObjective(BattleType.FORCE_COORDS, objectiveString) instanceof Location coords) {
+                double targetX = coords.getX();
+                double targetZ = coords.getZ();
+
+                double distance = player.getLocation().distance(
+                        new Location(player.getWorld(), targetX, player.getLocation().getY(), targetZ)
+                );
+
+                if (distance <= 2.0) {
+                    battlePlayer.sendMessage(Caption.of(
+                            "notification.objective_finished",
+                            TagResolver.resolver("objective", Tag.inserting(Component.text(targetX + ", " + targetZ)))
+                    ));
+                    Toast.display(
+                            player,
+                            Material.COMPASS,
+                            ChatColor.BLUE + Format.formatName(objectiveString),
+                            Toast.Style.GOAL,
+                            ForceBattle.get()
+                    );
+                    battlePlayer.updateObjective(true, false);
+                }
+            }
+        }
+    }
+
+}
