+++ src/main/java/net/fameless/forcebattle/game/tasks/ForceTask.java
@@ -1,0 +1,5 @@
+package net.fameless.forcebattle.game.tasks;
+
+public interface ForceTask {
+    void runTick();
+}
