+++ src/main/java/net/fameless/forcebattle/gui/ResultGUI.java
@@ -6,14 +6,14 @@ import net.fameless.forcebattle.caption.Caption;
 import net.fameless.forcebattle.game.Objective;
 import net.fameless.forcebattle.game.Team;
 import net.fameless.forcebattle.player.BattlePlayer;
-import net.fameless.forcebattle.util.Advancement;
+import net.fameless.forcebattle.game.data.Advancement;
 import net.fameless.forcebattle.util.BukkitUtil;
 import net.fameless.forcebattle.util.Format;
 import net.fameless.forcebattle.util.InventoryUtils;
 import net.fameless.forcebattle.util.ItemData;
 import net.fameless.forcebattle.util.ItemUtils;
 import net.fameless.forcebattle.util.ResultUtils;
-import net.fameless.forcebattle.util.Structure;
+import net.fameless.forcebattle.game.data.Structure;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.minimessage.tag.Tag;
 import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver;
